requirements.txt
python-whois
dnspython
requests
streamlit
streamlit-extras
jellyfish
numpy
email-validator
pyspf
dkimpy
dmarc
aiosmtplib
aiosqlite